+++ app/services/models.py
@@ -10,3 +10,11 @@ class Driver(BaseModel):
     name_acronym: str = Field(...)
     team_colour: str = Field(...)
     team_name: str = Field(...)
+
+
+class Location(BaseModel):
+    year: int = Field(...)
+    meeting_key: int = Field(...)
+    meeting_name: str = Field(...)
+    location: str = Field(...)
+    date_start: str = Field(...)   # Query need to be sorted by date_start